rename BulkPullReviews to PullsByRepo
